refactor(hub): deprecated alias barrels grouped under src/legacy/
Move the four deprecated-alias barrels (adapter, kernel, bundle, and
the new explicit describe stub) into src/legacy/, so they read as one
group of "old name, kept for pins" files instead of being scattered
across src/kernel/adapter/, src/kernel/index.ts, and src/bundle/.

/describe used to be a dual tsup entry pointing at the /ui source file;
it's now its own barrel (src/legacy/describe.ts) re-exporting /ui, so
the alias is visible code instead of a build-config side effect.
Repoint with-cargo/index.ts's `export * from '../kernel/index.js'`
floor and the surface-golden tests to the new paths. Entry names and
package.json exports are unchanged, so dist layout and published
subpaths are untouched; all golden baselines are byte-identical.

Also drop the now-dead kernel/index.ts grandfather entry from
PRE_EXISTING_SPINE_SERVICE_IMPORTS and the kernel/adapter/ exclusion
in port-layering's spine walk — both were made obsolete by this move.
