+++ packages/hub/__tests__/create-owner-adopted-partition.test.ts
@@ -147,7 +147,7 @@ describe('full ceremony end-to-end', () => {
 
 describe('bundle subpath export', () => {
   it('is exported from the @noy-db/hub/bundle subpath', async () => {
-    const mod = await import('../src/bundle/index.js')
+    const mod = await import('../src/legacy/bundle.js')
     expect(typeof mod.createOwnerOnAdoptedPartition).toBe('function')
   })
 })
