+++ packages/hub/src/legacy/kernel.ts
@@ -8,15 +8,18 @@
  * orchestration layer needs from the vault core. Treat it as a
  * contract — additive changes only; removals are breaking.
  *
+ * @deprecated Prefer `@noy-db/hub/cargo`, which carries this whole surface
+ * plus the orchestration delta. Kept for existing pins.
+ *
  * @packageDocumentation
  */
 
 // ─── runtime helpers ──────────────────────────────────────────────
-export { readPath } from './query/predicate.js'
+export { readPath } from '../kernel/query/predicate.js'
 export { reduceRecords } from '../with-lookup/aggregate/aggregation.js'
 export { groupAndReduce } from '../with-lookup/aggregate/groupby.js'
 export { generateULID } from '../with-pod/ulid.js'
-export { sha256Hex } from './enclave/index.js'
+export { sha256Hex } from '../kernel/enclave/index.js'
 // Coordination port — the stable drain-barrier seam an outward
 // orchestrator (@klum-db/lobby) or a `by-*` transport binds to. The pure
 // helpers are runtime; the port + presence/fence shapes are types only
@@ -37,7 +40,7 @@ export {
   UnknownShardError,
   ValidationError,
   VaultTemplateNotFoundError,
-} from './errors.js'
+} from '../kernel/errors.js'
 
 // Types only (erased at emit). NOTE: Vault / Collection / Noydb / Query are
 // runtime classes in hub, but are re-exported here as TYPES — `instanceof`
@@ -45,14 +48,14 @@ export {
 // runtime class value must import it from `@noy-db/hub` directly.
 // ─── types ────────────────────────────────────────────────────────
 export type { CollectionMeta, VaultMeta } from '../with-shape/introspection/meta.js'
-export type { ChangeEvent } from './types.js'
-export type { Vault } from './vault.js'
-export type { Collection } from './collection.js'
-export type { Noydb } from './noydb.js'
-export type { Operator } from './query/predicate.js'
-export type { Query } from './query/builder.js'
-export type { JoinStrategy } from './query/join.js'
-export type { LiveQuery } from './query/live.js'
+export type { ChangeEvent } from '../kernel/types.js'
+export type { Vault } from '../kernel/vault.js'
+export type { Collection } from '../kernel/collection.js'
+export type { Noydb } from '../kernel/noydb.js'
+export type { Operator } from '../kernel/query/predicate.js'
+export type { Query } from '../kernel/query/builder.js'
+export type { JoinStrategy } from '../kernel/query/join.js'
+export type { LiveQuery } from '../kernel/query/live.js'
 export type {
   AggregateResult,
   AggregateSpec,
