+++ packages/hub/src/with-cargo/README.md
@@ -9,7 +9,7 @@ orchestration delta klum previously pulled from the bare `@noy-db/hub` root
 barrel, so an outward orchestrator binds **one** narrow subpath instead of
 reaching into hub internals.
 
-`/kernel` (`src/kernel/index.ts`) remains as a **deprecated alias** for existing
+`/kernel` (`src/legacy/kernel.ts`) remains as a **deprecated alias** for existing
 pins and will **not be removed** without a coordinated version bump. New
 orchestration consumers should bind `/cargo`. The export surface is frozen by
 `__tests__/cargo-surface-golden.test.ts` against `cargo-surface.golden.json`, so
