+++ packages/hub/tsup.config.ts
@@ -31,7 +31,7 @@ const ENTRIES = {
   'indexing/index': 'src/with-lookup/indexing/index.ts',
   'aggregate/index': 'src/with-lookup/aggregate/index.ts',
   'crdt/index': 'src/with-commit/crdt/index.ts',
-  'bundle/index': 'src/bundle/index.ts',
+  'bundle/index': 'src/legacy/bundle.ts',
   'pod/index': 'src/with-pod/index.ts',
   'consent/index': 'src/with-audit/consent/index.ts',
   'periods/index': 'src/with-audit/periods/index.ts',
@@ -47,15 +47,16 @@ const ENTRIES = {
   'attestation/index': 'src/with-audit/attestation/index.ts',
   'tiers/index': 'src/with-audit/tiers/index.ts',
   'portability/index': 'src/with-audit/portability/index.ts',
-  'kernel/index': 'src/kernel/index.ts',
+  'kernel/index': 'src/legacy/kernel.ts',
   'cargo/index': 'src/with-cargo/index.ts',
-  'adapter/index': 'src/kernel/adapter/index.ts',
+  'adapter/index': 'src/legacy/adapter.ts',
   'to/index': 'src/port/to/index.ts',
   'with/index': 'src/port/with/index.ts',
   'ui/index': 'src/port/ui/index.ts',
-  // Deprecated alias: same source as 'ui/index', kept so `dist/describe/index.js`
-  // (the pre-S5 `./describe` subpath) continues to exist. See port/ui/index.ts header.
-  'describe/index': 'src/port/ui/index.ts',
+  // Deprecated alias: re-exports the 'ui/index' surface, kept so
+  // `dist/describe/index.js` (the pre-S5 `./describe` subpath) continues
+  // to exist. See src/legacy/describe.ts.
+  'describe/index': 'src/legacy/describe.ts',
   'by/index': 'src/port/by/index.ts',
   'on/index': 'src/port/on/index.ts',
   'at/index': 'src/port/at/index.ts',
